Make tool more scalable + prettify report
* Two new modules were added: Providers and Reports. Built on top of Core,
that now only contains common/shared stuff and Apple framework's extensions.

* Subcommands were added, including the default FullAnalyzes, to allow the user
to generate not only a full report, but one for a specific provider.
Such change itends to make it easier to support the addition of arguments
there are specific to subcommands in the future, if needed.
As another solution, ff it doesn't work well as expected, things can be moved back
to a single command with the addition of e.g. an argument set to define
which providers to use when generating a report.