Clone app in current dir
* This fixes adding local package as dependency,
by cloning the measurement app into current directory
instead of in a temporary one, which is needed for the
relative path to work correctly. As side effect this brings the
risk of lefting the user directory dirty in case a user
manually cancels the tool after the empty app was cloned.